backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
backboard-sdk>=1.4.0
motor==3.3.2
pymongo==4.6.1
snowflake-connector-python
